pyproject.toml
[project]
name = "pyml-regression-example1"
version = "0.1.0"
description = "Python Machine Learning Regression Example: Life Expectancy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "colorama>=0.4.6",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "pandas>=2.2.3",
    "platformdirs>=4.3.6",
    "scikit-learn>=1.5.2",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-click-rst-to-ansi-formatter>=0.1.2",
]
packages = [{include = "life_expectancy", from = "src"}]

[project.scripts]
life-expectancy = "life_expectancy.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/life_expectancy"]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "mypy>=1.11.2",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
    "ruff>=0.6.6",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx>=7.0.0",
    "sphinx-click>=6.0.0",
    "types-click>=7.1.8",
    "types-colorama>=0.4.15.20240311",
    "pre-commit>=3.8.0",
    "rstcheck>=6.2.4",
    "tox>=4.20.0",
    "types-requests>=2.32.0.20240914",
    "matplotlib-stubs>=0.1.0",
    "pandas-stubs>=2.2.2.240909",
]

[tool.coverage.report]
fail_under = 100
